修改：恢复Welcome布局，重构Login
