+++ app/src/main/AndroidManifest.xml
@@ -45,6 +45,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".AddCommentActivity" />
         <activity android:name=".RegActivity" />
+        <activity android:name=".LoginActivity"></activity>
     </application>
 
 </manifest>
