+++ app/src/main/res/layout/activity_welcome.xml
@@ -1,78 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
+    android:background="@color/colorPrimary"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="cn.edu.xjtu.se.bookgamma.WelcomeActivity">
 
     <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="250dp"
-        android:layout_gravity="center_horizontal"
-        android:gravity="center_vertical"
-        android:text="BookGamma"
-        android:textSize="36sp" />
-
-    <android.support.design.widget.TextInputLayout
-        style="@style/Style.Margin.Normal"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="10dp"
-        android:layout_marginLeft="30dp"
-        android:layout_marginRight="30dp">
+        android:layout_height="match_parent"
+        android:gravity="center"
+        android:textSize="36sp"
+        android:textColor="@color/primary_light"
+        android:text="BookGamma\n\nWelcome" />
 
-        <EditText
-            android:id="@+id/et_user"
-            android:layout_width="match_parent"
-            android:focusableInTouchMode="false"
-            android:layout_height="wrap_content"
-            android:hint="@string/tx_username"
-            android:textSize="16sp" />
-    </android.support.design.widget.TextInputLayout>
-
-    <android.support.design.widget.TextInputLayout
-        style="@style/Style.Margin.Normal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="30dp"
-        android:layout_marginRight="30dp">
-
-        <EditText
-            android:id="@+id/et_password"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/tx_password"
-            android:inputType="textPassword"
-            android:textSize="16sp" />
-    </android.support.design.widget.TextInputLayout>
-
-    <Button
-        android:id="@+id/btn_login"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="30dp"
-        android:layout_marginRight="30dp"
-        android:layout_marginTop="20dp"
-        android:background="@color/colorPrimary"
-        android:text="@string/btn_login"
-        android:textColor="@android:color/white"
-        android:textSize="18sp" />
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <Button
-            android:id="@+id/btn_register"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentRight="true"
-            android:background="@null"
-            android:padding="30dp"
-            android:text="@string/btn_register"
-            android:textColor="@color/colorPrimary"
-            android:textSize="18sp" />
-    </RelativeLayout>
-</LinearLayout>
+</RelativeLayout>
